Move code for CSR into extension
